Rename syntaxKindVisit file
