+++ src/accept/syntaxKindVisit.ts
